requirements.txt
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
pygame==1.9.3
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
